debug modal rework into comp api
