+++ web/src/components/ui/DialogWrapper.vue
@@ -1,16 +1,14 @@
 <template>
-  <q-dialog ref="dialog" @hide="onHide">
-    <q-card class="q-dialog-plugin">
+  <q-dialog ref="dialog" @hide="onHide" v-bind="dialogProps">
+    <q-card class="q-dialog-plugin" :style="`min-width: ${width}vw`">
       <q-bar>
-        <slot name="bar">
-          {{ title }}
-          <q-space />
-          <q-btn dense flat icon="close" v-close-popup>
-            <q-tooltip class="bg-white text-primary">Close</q-tooltip>
-          </q-btn>
-        </slot>
+        {{ title }}
+        <q-space />
+        <q-btn dense flat icon="close" v-close-popup>
+          <q-tooltip content-class="bg-white text-primary">Close</q-tooltip>
+        </q-btn>
       </q-bar>
-      <div class="scroll" style="height: 70vh">
+      <div class="scroll" :style="`height: ${height}vh`">
         <component :is="vuecomponent" v-bind="{ ...$attrs, ...componentProps }" @close="onOk" @hide="hide" />
       </div>
     </q-card>
@@ -24,7 +22,16 @@ export default {
   props: {
     vuecomponent: {},
     title: String,
+    width: {
+      type: String,
+      default: "50",
+    },
+    height: {
+      type: String,
+      default: "50",
+    },
     componentProps: Object,
+    dialogProps: Object,
   },
   inheritAttrs: false,
   methods: {
